requirements.txt
apscheduler==3.10.0
asyncpg==0.27.0
inflect==6.0.2
pymorphy2==0.9.1
pymorphy2-dicts-uk==2.4.1.1.1460299261
pyrogram==2.0.99
python-dateutil==2.8.2
simplejson==3.18.3
sqlalchemy==2.0.4
tgcrypto==1.2.5